+++ src/utils/collection.js
@@ -1,6 +1,7 @@
 import { isCollection, isMatrix } from './is.js'
 import { IndexError } from '../error/IndexError.js'
 import { arraySize } from './array.js'
+import { _switch } from './switch.js'
 
 /**
  * Test whether an array contains collections
@@ -119,27 +120,6 @@ function _reduce (mat, dim, callback) {
     }
     return ret
   }
-}
-
-/**
- * Transpose a matrix
- * @param {Array} mat
- * @returns {Array} ret
- * @private
- */
-function _switch (mat) {
-  const I = mat.length
-  const J = mat[0].length
-  let i, j
-  const ret = []
-  for (j = 0; j < J; j++) {
-    const tmp = []
-    for (i = 0; i < I; i++) {
-      tmp.push(mat[i][j])
-    }
-    ret.push(tmp)
-  }
-  return ret
 }
 
 // TODO: document function scatter